Move TLSv1.0 and SSLv3 deprecation tests to a separate class in platform/
